+++ app/behaviourspc/FallBackFlow.scala
@@ -16,21 +16,18 @@
 
 package behaviourspc
 
-import flow.InitialBehaviour
 import models.InteractionCategories.{CardReader, OnlineCategory}
 import models.TranResults.SuccessResult
 import models._
 
-class StandardInitialBehaviour(flowData: SpcFlowData) extends InitialBehaviour {
+class FallBackFlow(spcFlow: FlowData) extends Flow {
 
   val initialBehaviour: SpcBehaviour = handlePedLogOn
 
   private lazy val handlePedLogOn: SpcBehaviour = behave {
     case pedLogOn: PedLogOn =>
-      (
-        List(PedLogOnResponse(HeaderNode(), pedLogOn.messageNode, SuccessResult, ErrorsNode(Seq.empty))),
-        handleSubmitPayment orElse CommonBehaviours.handlePedLogOff
-      )
+      val pedLogOnResponse: SpcResponseMessage = PedLogOnResponse(HeaderNode(), pedLogOn.messageNode, SuccessResult, ErrorsNode(Seq.empty))
+      (List(pedLogOnResponse), handleSubmitPayment orElse CommonBehaviours.handlePedLogOff)
   }
 
   private lazy val handleSubmitPayment: SpcBehaviour = behave {
@@ -44,19 +41,15 @@ class StandardInitialBehaviour(flowData: SpcFlowData) extends InitialBehaviour {
       )
 
       val submitPaymentResponse = SubmitPaymentResponse(HeaderNode(), submitPayment.messageNode, SuccessResult)
-      (
-        List(submitPaymentResponse),
-        handleProcessTransaction(paymentSubmittedData)
-      )
+      (List(submitPaymentResponse), handleProcessTransaction(paymentSubmittedData))
   }
 
   //sends UpdatePaymentEnhanced
-  private def handleProcessTransaction(submittedData: SubmittedData): SpcBehaviour = behave{
-
+  private def handleProcessTransaction(submittedData: SubmittedData): SpcBehaviour = behave {
     case processTransaction: ProcessTransaction =>
 
       //Display sequence - card validation
-      val interimResponses = flowData.displayMessagesValidation.map {
+      val interimResponses = spcFlow.displayMessagesValidation.map{
         case (interactionEvents, interactionPrompt) =>
           val interactionNode = InteractionNode(category = CardReader, event = interactionEvents, prompt = interactionPrompt)
           val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), processTransaction.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
@@ -66,13 +59,11 @@ class StandardInitialBehaviour(flowData: SpcFlowData) extends InitialBehaviour {
       //UpdatePaymentEnhanced
       val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
       val transactionNode = TransactionNode(amountNode = amountNode)
-      val cardNode = UpeCardNode(flowData.paymentCard)
+      val cardNode = UpeCardNode(spcFlow.paymentCard)
       val updatePaymentEnhanced = UpdatePaymentEnhanced(HeaderNode(), processTransaction.messageNode, transactionNode, cardNode, SuccessResult, ErrorsNode(Seq.empty))
-      val responses: Seq[SpcResponseMessage] = interimResponses :+[SpcResponseMessage] updatePaymentEnhanced
-
       (
-        responses,
-        handleUpdatePaymentEnhancedResponse(submittedData) orElse handleTransactionCancelled(submittedData)
+        interimResponses :+[SpcResponseMessage] updatePaymentEnhanced,
+        handleUpdatePaymentEnhancedResponse(submittedData) orElse handleTransactionCancelled(submittedData) orElse handleGetTransactionDetails(submittedData)
       )
   }
 
@@ -82,7 +73,7 @@ class StandardInitialBehaviour(flowData: SpcFlowData) extends InitialBehaviour {
       val totalAmount = updatePaymentEnhancedResponse.amountNode.totalAmount
 
       //Display sequence - card Authentication
-      val interimResponses = flowData.displayMessagesAuthentication.map{
+      val interimResponses = spcFlow.displayMessagesAuthentication.map {
         case (interactionEvents, interactionPrompt) =>
           val interactionNode = InteractionNode(category = OnlineCategory, event = interactionEvents, prompt = interactionPrompt)
           val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), updatePaymentEnhancedResponse.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
@@ -90,59 +81,18 @@ class StandardInitialBehaviour(flowData: SpcFlowData) extends InitialBehaviour {
       }
 
       //PosPrintReceipt client
-      val merchantReceiptNode = ReceiptMerchantNode(flowData, submittedData, totalAmount, finalAmount)
+      val merchantReceiptNode = ReceiptMerchantNode(spcFlow, submittedData, totalAmount, finalAmount)
       val posPrintReceipt = PosPrintReceipt(HeaderNode(), updatePaymentEnhancedResponse.messageNode, merchantReceiptNode, SuccessResult, ErrorsNode(Seq.empty))
-      val responses: Seq[SpcResponseMessage] = interimResponses :+[SpcResponseMessage] posPrintReceipt
 
+      val posDecisionTransNode = PdTransNode(TransactionDecisions.SignatureRequired)
+      val posDecisionMessage = PosDecisionMessage(HeaderNode(), updatePaymentEnhancedResponse.messageNode, posDecisionTransNode)
       (
-        responses,
-        handlePosPrintReceiptResponse(submittedData, totalAmount, finalAmount, merchantReceiptNode)
+        interimResponses :+[SpcResponseMessage] posPrintReceipt :+ posDecisionMessage,
+        handleTransactionCancelled(submittedData) orElse handleGetTransactionDetails(submittedData)
       )
   }
 
-  private def handlePosPrintReceiptResponse(submittedData: SubmittedData, totalAmount: AmountInPence, finalAmount: Option[AmountInPence], merchantReceiptNode: ReceiptNode): SpcBehaviour = behave{
-    case posPrintReceiptResponse: PosPrintReceiptResponse =>
-
-      //PosPrintReceipt client
-      val clientReceiptNode = ReceiptNode.createReceiptNode(submittedData, flowData, totalAmount, finalAmount)
-
-      val posPrintReceipt = PosPrintReceipt(HeaderNode(), posPrintReceiptResponse.messageNode, clientReceiptNode, SuccessResult, ErrorsNode(Seq.empty))
-      (
-        List(posPrintReceipt),
-        handlePosPrintReceiptResponseWithPtr(submittedData, totalAmount, finalAmount, merchantReceiptNode, clientReceiptNode)
-      )
-  }
-
-  private def handlePosPrintReceiptResponseWithPtr(submittedData: SubmittedData, totalAmount: AmountInPence, finalAmount: Option[AmountInPence], merchantReceiptNode: ReceiptNode, clientReceiptNode: ReceiptNode): SpcBehaviour = behave{
-    case posPrintReceiptResponse: PosPrintReceiptResponse =>
-
-      //processTransactionResponse
-      val amountNode = AmountNode(totalAmount, submittedData.currency, submittedData.country, finalAmount)
-
-      val ptrTransactionNode = PtrTransactionNode(
-        amountNode      = amountNode,
-        verification    = flowData.cardVerificationMethod,
-        transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
-        transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
-      val cardNode = PtrResponseCardNode(flowData.paymentCard)
-
-      val processTransactionResponse = ProcessTransactionResponse(
-        headerNode           = HeaderNode(),
-        messageNode          = posPrintReceiptResponse.messageNode,
-        ptrTransactionNode   = ptrTransactionNode,
-        ptrCardNode          = cardNode,
-        result               = flowData.transactionResult,
-        paymentResult        = flowData.paymentResult,
-        receiptNodeCustomerO = Some(clientReceiptNode),
-        receiptNodeMerchantO = Some(merchantReceiptNode),
-        errorsNode           = ErrorsNode(Seq.empty))
-
-      (List(processTransactionResponse),
-        CommonBehaviours.handleFinalise
-      )
-  }
-
-  private def handleTransactionCancelled(submittedData: SubmittedData): SpcBehaviour = behave{
+  private def handleTransactionCancelled(submittedData: SubmittedData): SpcBehaviour = behave {
     case cancelTransaction: CancelTransaction =>
 
       //processTransactionResponse
@@ -150,26 +100,47 @@ class StandardInitialBehaviour(flowData: SpcFlowData) extends InitialBehaviour {
 
       val ptrTransactionNode = PtrTransactionNode(
         amountNode      = amountNode,
-        verification    = flowData.cardVerificationMethod,
+        verification    = spcFlow.cardVerificationMethod,
         transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
         transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
-      val cardNode = PtrResponseCardNode(flowData.paymentCard)
+      val cardNode = PtrResponseCardNode(spcFlow.paymentCard)
 
       val processTransactionResponse = ProcessTransactionResponse(
         headerNode           = HeaderNode(),
         messageNode          = cancelTransaction.messageNode,
         ptrTransactionNode   = ptrTransactionNode,
         ptrCardNode          = cardNode,
-        result               = flowData.transactionResult,
+        result               = spcFlow.transactionResult,
         paymentResult        = PaymentResults.cancelled,
         receiptNodeCustomerO = None,
         receiptNodeMerchantO = None,
         errorsNode           = ErrorsNode(Seq.empty))
-
-      (List(processTransactionResponse),
-        CommonBehaviours.handleFinalise
-      )
+      (List(processTransactionResponse), CommonBehaviours.handleFinalise)
   }
 
+  private def handleGetTransactionDetails(submittedData: SubmittedData): SpcBehaviour = behave {
+    case getTransactionDetails: GetTransactionDetails =>
+      //processTransactionResponse
+      val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
+
+      val ptrTransactionNode = PtrTransactionNode(
+        amountNode      = amountNode,
+        verification    = spcFlow.cardVerificationMethod,
+        transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
+        transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
+      val cardNode = PtrResponseCardNode(spcFlow.paymentCard)
+
+      val processTransactionResponse = GetTransactionDetailsResponse(
+        headerNode           = HeaderNode(),
+        messageNode          = getTransactionDetails.messageNode,
+        ptrTransactionNode   = ptrTransactionNode,
+        ptrCardNode          = cardNode,
+        result               = spcFlow.transactionResult,
+        paymentResult        = PaymentResults.cancelled,
+        receiptNodeCustomerO = None,
+        receiptNodeMerchantO = None,
+        errorsNode           = ErrorsNode(Seq.empty))
+      (List(processTransactionResponse), CommonBehaviours.handleFinalise)
+  }
 
 }
