+++ app/behaviourspc/NoReceiptFlow.scala
@@ -16,17 +16,17 @@
 
 package behaviourspc
 
-import flow.InitialBehaviour
-import models.InteractionCategories.CardReader
+import models.InteractionCategories.{CardReader, OnlineCategory}
 import models.TranResults.SuccessResult
 import models._
 
-class BinCheckCardDiscardedInitialBehaviour(spcFlow: SpcFlowData) extends InitialBehaviour {
+class NoReceiptFlow(flowData: FlowDataNoReceipt) extends Flow {
 
   val initialBehaviour: SpcBehaviour = handlePedLogOn
 
-  private lazy val  handlePedLogOn: SpcBehaviour = behave {
+  private lazy val handlePedLogOn: SpcBehaviour = behave {
     case pedLogOn: PedLogOn =>
+
       val pedLogOnResponse: SpcResponseMessage = PedLogOnResponse(HeaderNode(), pedLogOn.messageNode, SuccessResult, ErrorsNode(Seq.empty))
       (
         List(pedLogOnResponse),
@@ -34,7 +34,7 @@ class BinCheckCardDiscardedInitialBehaviour(spcFlow: SpcFlowData) extends Initia
       )
   }
 
-  private lazy val  handleSubmitPayment: SpcBehaviour = behave{
+  private def handleSubmitPayment: SpcBehaviour = behave {
     case submitPayment: SubmitPayment =>
       val paymentSubmittedData = SubmittedData(
         totalAmount         = submitPayment.transactionNode.amountNode.totalAmount,
@@ -45,17 +45,14 @@ class BinCheckCardDiscardedInitialBehaviour(spcFlow: SpcFlowData) extends Initia
       )
 
       val submitPaymentResponse = SubmitPaymentResponse(HeaderNode(), submitPayment.messageNode, SuccessResult)
-      (
-        List(submitPaymentResponse),
-        handleProcessTransaction(paymentSubmittedData)
-      )
+      (List(submitPaymentResponse), handleProcessTransaction(paymentSubmittedData))
   }
 
   //sends UpdatePaymentEnhanced
   private def handleProcessTransaction(submittedData: SubmittedData): SpcBehaviour = behave {
     case processTransaction: ProcessTransaction =>
       //Display sequence - card validation
-      val interimResponses: Seq[PosDisplayMessage] = spcFlow.displayMessagesValidation.map{
+      val interimMessages = flowData.displayMessagesValidation.map{
         case (interactionEvents, interactionPrompt) =>
           val interactionNode = InteractionNode(category = CardReader, event = interactionEvents, prompt = interactionPrompt)
           val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), processTransaction.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
@@ -65,37 +62,71 @@ class BinCheckCardDiscardedInitialBehaviour(spcFlow: SpcFlowData) extends Initia
       //UpdatePaymentEnhanced
       val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
       val transactionNode = TransactionNode(amountNode = amountNode)
-      val cardNode = UpeCardNode(spcFlow.paymentCard)
+      val cardNode = UpeCardNode(flowData.paymentCard)
       val updatePaymentEnhanced = UpdatePaymentEnhanced(HeaderNode(), processTransaction.messageNode, transactionNode, cardNode, SuccessResult, ErrorsNode(Seq.empty))
+      (interimMessages :+[SpcResponseMessage] updatePaymentEnhanced, handleUpdatePaymentEnhancedResponse(submittedData) orElse handleTransactionCancelled(submittedData))
+  }
 
-      (
-        interimResponses :+[SpcResponseMessage] updatePaymentEnhanced,
-        handleTransactionCancelled(submittedData)
-      )
+  private def handleUpdatePaymentEnhancedResponse(submittedData: SubmittedData): SpcBehaviour = behave {
+    case updatePaymentEnhancedResponse: UpdatePaymentEnhancedResponse =>
+      val finalAmount = updatePaymentEnhancedResponse.amountNode.finalAmountO
+      val totalAmount = updatePaymentEnhancedResponse.amountNode.totalAmount
+
+      //Display sequence - card Authentication
+      val interimMessages = flowData.displayMessagesAuthentication.map{
+        case (interactionEvents, interactionPrompt) =>
+          val interactionNode = InteractionNode(category = OnlineCategory, event = interactionEvents, prompt = interactionPrompt)
+          val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), updatePaymentEnhancedResponse.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
+          posDisplayMessageInsertCard
+      }
+
+      //processTransactionResponse
+      val amountNode = AmountNode(totalAmount, submittedData.currency, submittedData.country, finalAmount)
+
+      val ptrTransactionNode = PtrTransactionNode(
+        amountNode      = amountNode,
+        verification    = flowData.cardVerificationMethod,
+        transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
+        transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
+      val cardNode = PtrResponseCardNode(flowData.paymentCard)
+
+      val processTransactionResponse = ProcessTransactionResponse(
+        headerNode           = HeaderNode(),
+        messageNode          = updatePaymentEnhancedResponse.messageNode,
+        ptrTransactionNode   = ptrTransactionNode,
+        ptrCardNode          = cardNode,
+        result               = flowData.transactionResult,
+        paymentResult        = flowData.paymentResult,
+        receiptNodeCustomerO = None,
+        receiptNodeMerchantO = None,
+        errorsNode           = ErrorsNode(Seq.empty))
+      (interimMessages :+[SpcResponseMessage] processTransactionResponse, CommonBehaviours.handleFinalise)
   }
 
   private def handleTransactionCancelled(submittedData: SubmittedData): SpcBehaviour = behave {
     case cancelTransaction: CancelTransaction =>
+
       //processTransactionResponse
       val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
 
       val ptrTransactionNode = PtrTransactionNode(
         amountNode      = amountNode,
-        verification    = spcFlow.cardVerificationMethod,
+        verification    = flowData.cardVerificationMethod,
         transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
         transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
-      val cardNode = PtrResponseCardNode(spcFlow.paymentCard)
+      val cardNode = PtrResponseCardNode(flowData.paymentCard)
 
       val processTransactionResponse = ProcessTransactionResponse(
         headerNode           = HeaderNode(),
         messageNode          = cancelTransaction.messageNode,
         ptrTransactionNode   = ptrTransactionNode,
         ptrCardNode          = cardNode,
-        result               = spcFlow.transactionResult,
+        result               = flowData.transactionResult,
         paymentResult        = PaymentResults.cancelled,
         receiptNodeCustomerO = None,
         receiptNodeMerchantO = None,
         errorsNode           = ErrorsNode(Seq.empty))
+
       (List(processTransactionResponse), CommonBehaviours.handleFinalise)
   }
