+++ app/behaviourspc/PedDisconnectedFlow.scala
@@ -16,22 +16,24 @@
 
 package behaviourspc
 
-import flow.InitialBehaviour
-import models.InteractionCategories.CardReader
 import models.TranResults.SuccessResult
 import models._
 
-class NoSurchargeInitialBehaviour(spcFlow: SpcFlowData, errorsNode: ErrorsNode) extends InitialBehaviour {
+class PedDisconnectedFlow(spcFlow: FlowDataNoReceipt, errorsNode: ErrorsNode) extends Flow {
 
   val initialBehaviour: SpcBehaviour = handlePedLogOn
 
   private lazy val handlePedLogOn: SpcBehaviour = behave {
     case pedLogOn: PedLogOn =>
+
       val pedLogOnResponse: SpcResponseMessage = PedLogOnResponse(HeaderNode(), pedLogOn.messageNode, SuccessResult, ErrorsNode(Seq.empty))
-      (List(pedLogOnResponse), handleSubmitPayment orElse CommonBehaviours.handlePedLogOff)
+      (List(
+        pedLogOnResponse),
+        handleSubmitPayment orElse CommonBehaviours.handlePedLogOff
+      )
   }
 
-  private lazy val  handleSubmitPayment: SpcBehaviour = behave {
+  private lazy val handleSubmitPayment: SpcBehaviour = behave {
     case submitPayment: SubmitPayment =>
       val paymentSubmittedData = SubmittedData(
         totalAmount         = submitPayment.transactionNode.amountNode.totalAmount,
@@ -42,49 +44,18 @@ class NoSurchargeInitialBehaviour(spcFlow: SpcFlowData, errorsNode: ErrorsNode) 
       )
 
       val submitPaymentResponse = SubmitPaymentResponse(HeaderNode(), submitPayment.messageNode, SuccessResult)
-      (List(submitPaymentResponse), handleProcessTransaction(paymentSubmittedData))
+      (
+        List(submitPaymentResponse),
+        handleProcessTransaction(paymentSubmittedData) orElse handleTransactionCancelled(paymentSubmittedData)
+      )
   }
 
   //Do not send updatePaymentEnhanced but got to print message instead
   private def handleProcessTransaction(submittedData: SubmittedData): SpcBehaviour = behave {
 
     case processTransaction: ProcessTransaction =>
-
-      //Display sequence - card validation
-      val interimMessages = spcFlow.displayMessagesValidation.map{
-        case (interactionEvents, interactionPrompt) =>
-          val interactionNode = InteractionNode(category = CardReader, event = interactionEvents, prompt = interactionPrompt)
-          val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), processTransaction.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
-          posDisplayMessageInsertCard
-      }
-
-      //PosPrintReceipt client
-      val merchantReceiptNode = ReceiptMerchantNode(spcFlow, submittedData, submittedData.totalAmount, None)
-      val posPrintReceipt: PosPrintReceipt = PosPrintReceipt(HeaderNode(), processTransaction.messageNode, merchantReceiptNode, SuccessResult, ErrorsNode(Seq.empty))
-      (
-        interimMessages :+[SpcResponseMessage] posPrintReceipt,
-        handlePosPrintReceiptResponse(submittedData, None, merchantReceiptNode) orElse handleTransactionCancelled(submittedData)
-      )
-  }
-
-  private def handlePosPrintReceiptResponse(submittedData: SubmittedData, finalAmount: Option[AmountInPence], merchantReceiptNode: ReceiptNode): SpcBehaviour = behave {
-    case posPrintReceiptResponse: PosPrintReceiptResponse =>
-
-      //PosPrintReceipt client
-      val clientReceiptNode = ReceiptNode.createReceiptNode(submittedData, spcFlow, submittedData.totalAmount, finalAmount)
-
-      val posPrintReceipt = PosPrintReceipt(HeaderNode(), posPrintReceiptResponse.messageNode, clientReceiptNode, SuccessResult, ErrorsNode(Seq.empty))
-      (
-        List(posPrintReceipt),
-        handlePosPrintReceiptResponseWithPtr(submittedData, finalAmount, merchantReceiptNode, clientReceiptNode)
-      )
-
-  }
-
-  private def handlePosPrintReceiptResponseWithPtr(submittedData: SubmittedData, finalAmount: Option[AmountInPence], merchantReceiptNode: ReceiptNode, clientReceiptNode: ReceiptNode): SpcBehaviour = behave {
-    case posPrintReceiptResponse: PosPrintReceiptResponse =>
       //processTransactionResponse
-      val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, finalAmount)
+      val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
 
       val ptrTransactionNode = PtrTransactionNode(
         amountNode      = amountNode,
@@ -95,15 +66,18 @@ class NoSurchargeInitialBehaviour(spcFlow: SpcFlowData, errorsNode: ErrorsNode) 
 
       val processTransactionResponse = ProcessTransactionResponse(
         headerNode           = HeaderNode(),
-        messageNode          = posPrintReceiptResponse.messageNode,
+        messageNode          = processTransaction.messageNode,
         ptrTransactionNode   = ptrTransactionNode,
         ptrCardNode          = cardNode,
         result               = spcFlow.transactionResult,
         paymentResult        = spcFlow.paymentResult,
-        receiptNodeCustomerO = Some(clientReceiptNode),
-        receiptNodeMerchantO = Some(merchantReceiptNode),
+        receiptNodeCustomerO = None,
+        receiptNodeMerchantO = None,
         errorsNode           = errorsNode)
-      (List(processTransactionResponse), CommonBehaviours.handleFinalise)
+      (
+        List(processTransactionResponse),
+        CommonBehaviours.handleFinalise
+      )
   }
 
   private def handleTransactionCancelled(submittedData: SubmittedData): SpcBehaviour = behave {
@@ -128,8 +102,10 @@ class NoSurchargeInitialBehaviour(spcFlow: SpcFlowData, errorsNode: ErrorsNode) 
         receiptNodeCustomerO = None,
         receiptNodeMerchantO = None,
         errorsNode           = ErrorsNode(Seq.empty))
-      (List(processTransactionResponse), CommonBehaviours.handleFinalise)
+      (
+        List(processTransactionResponse),
+        CommonBehaviours.handleFinalise
+      )
   }
-
 
 }
